- TorchManager, TempleSwitchManager -> EventManager로 병합 - Enter/Exit House 집 문쪽에 추가 - 보스 몬스터 죽을 시 EventManager 통해서 문열리도록 UAssasinBossFSMComponent::UpdateDying에 로직 추가 : OpenGateSequence 호출 시 모든 게이트가 열려서 다시 수정해야 함 아직 고민 중
