--- TeamProject/Source/TeamProject/SubSystem/Puzzle/TempleSwitchManager.cpp
@@ -1,9 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "SubSystem/Puzzle/TempleSwitchManager.h"
-
-void UTempleSwitchManager::NotifyOverlapTempleBall()
-{
-	OnOverlapTempleBall.Broadcast();
-}
